Removed getflagVariationByKey from optimizely and moved it to project config also instead of getting variation from rule
